moved views and functions related to Aggregation profiles to a separate file
